oindrieel/requirements.txt
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
pydantic